Lots of rethinking
Renaming some of the menu items to hopefully really clarify their purpose. Additionally, deciding what is shared and what is not. But realizing, too, that a shared action does not necessitate a shared implementation. Take New Tab, for example: it will have a base MenuModule action, but the implementation will soon vary between workspaces. Its command handler will be moved from FileService to Notepad and Notebook. In the former, it will behave as before (creating a new model and view onto a plain text document that exists only in memory). For the latter, though, it will likely create a new file inside the extracted archive, assign it a UUID, add it to the TreeView model, and then open it in a new tab (new model + view).