Fixing a very silly bug
Fixing the "models with 0 views not closing in Notepad" bug.

The issue was I was greatly overcomplicating it by choosing to emit the Bus::viewDestroyed signal manually everywhere EXCEPT the critical point (on actual destruction), which ultimately was the only place we needed to emit it at all.

Fixing this means much simpler ViewService closure methods, too.