Switched to normal iteration when gathering file views for saves
View closures are performed separately from saves, so making saves work in reverse order, too, is stupid and leads to some confusing outcomes from a user perspective (like--as in Notepad, at least not too long ago--when aborting a Save As dialog and seeing tabs ahead of that tab being saved (and closed) but not the tabs before it. I think as a user you'd expect it to work from left to right generally. So now it does.